+++ Libplanet.Tests/Action/PolymorphicActionTest.cs
@@ -112,6 +112,12 @@ namespace Libplanet.Tests.Action
             {
                 return context.PreviousStates;
             }
+
+            public void Render(
+                IActionContext context,
+                IAccountStateDelta nextStates)
+            {
+            }
         }
     }
 }