+++ Libplanet.Tests/Common/Action/ThrowException.cs
@@ -34,6 +34,12 @@ namespace Libplanet.Tests.Common.Action
             return context.PreviousStates;
         }
 
+        public void Render(
+            IActionContext context,
+            IAccountStateDelta nextStates)
+        {
+        }
+
         public class SomeException : Exception
         {
             public SomeException(string message)