+++ Libplanet/Action/PolymorphicAction.cs
@@ -46,6 +46,13 @@ namespace Libplanet.Action
     ///     // implement their own logic.
     ///     public abstract IAccountStateDelta Execute(IActionContext context);
     ///
+    ///     // Makes Render() no-op by default, but overrideable by subclasses.
+    ///     public virtual void Render(
+    ///         IActionContext context,
+    ///         IAccountStateDelta nextStates)
+    ///     {
+    ///     }
+    ///
     ///     IImmutableDictionary<string, object> IAction.PlainValue =>
     ///         ImmutableDictionary<string, object>.Empty
     ///             .Add("target_address", TargetAddress);
@@ -81,6 +88,19 @@ namespace Libplanet.Action
     ///             ImmutableDictionary<string, uint>.Empty.Add("hp", 20)
     ///         );
     ///     }
+    ///
+    ///     void IAction.Render(
+    ///         IActionContext context,
+    ///         IAccountStateDelta nextStates)
+    ///     {
+    ///         var c = new Character
+    ///         {
+    ///             Address = TargetAddress,
+    ///             Hp = nextStates.GetState(TargetAddress)["hp"],
+    ///         };
+    ///         c.Draw();
+    ///         break;
+    ///     }
     /// }
     ///
     /// [ActionType("attack")]
@@ -95,6 +115,15 @@ namespace Libplanet.Action
     ///             state.SetItem("hp", Math.Max(state["hp"] - 5, 0))
     ///         );
     ///     }
+    ///
+    ///     void IAction.Render(
+    ///         IActionContext context,
+    ///         IAccountStateDelta nextStates)
+    ///     {
+    ///         Character c = Character.GetByAddress(TargetAddress);
+    ///         c.Hp = nextStates.GetState(TargetAddress)["hp"];
+    ///         c.Draw();
+    ///     }
     /// }
     ///
     /// [ActionType("heal")]
@@ -108,6 +137,15 @@ namespace Libplanet.Action
     ///             TargetAddress,
     ///             state.SetItem("hp", Math.Min(state["hp"] + 5, 20))
     ///         );
+    ///     }
+    ///
+    ///     void IAction.Render(
+    ///         IActionContext context,
+    ///         IAccountStateDelta nextStates)
+    ///     {
+    ///         Character c = Character.GetByAddress(TargetAddress);
+    ///         c.Hp = nextStates.GetState(TargetAddress)["hp"];
+    ///         c.Draw();
     ///     }
     /// }
     /// ]]></code>
@@ -219,5 +257,12 @@ namespace Libplanet.Action
         /// <inheritdoc/>
         public IAccountStateDelta Execute(IActionContext context) =>
             InnerAction.Execute(context);
+
+        /// <inheritdoc/>
+        public void Render(
+            IActionContext context,
+            IAccountStateDelta nextStates
+        ) =>
+            InnerAction.Render(context, nextStates);
     }
 }